Get rid of remaining warnings

+++ core/src/main/c/windows/accept.c
@@ -30,7 +30,7 @@
 
 JNIEXPORT jint JNICALL Java_io_questdb_network_Net_abortAccept
         (JNIEnv *e, jclass cl, jint fd) {
-    jint r = CloseHandle((HANDLE) fd);
+    jint r = CloseHandle((HANDLE) ((uintptr_t) fd));
     if (!r) {
         SaveLastError();
     }
